MasterMind Game
This is based of the MasterMind Game.
The computer can run simulate multiple runs by changing
numberOfTimesPlayed.
Texted based game with ascii banner.
Enter colors in a certain order to guess the answer code.
When a guess is made, and answer is given as:
1 = correct color
2 = correct color and position
The numbers do not correlate with the positions of the colors.
Input one guess at a time in order from left to right.
Set to 6 colors and a guess size of 4 but can be changed to any number
in the BoardRunner.
The computer solver can only solve for even sets of colors right now.